pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "graph-fusion-experiment"
version = "1.0.0"
description = "Graph fusion pattern detection and optimization"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.8"
authors = [
    {name = "Sudheer Devu"}
]
keywords = ["graph", "fusion", "optimization", "compiler", "ai"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
]

dependencies = [
    "numpy>=1.21.0",
    "onnx>=1.12.0",
    "pyyaml>=6.0",
    "networkx>=2.6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=22.0.0",
    "isort>=5.10.0",
]

[project.urls]
Homepage = "https://github.com/sudheerdevu/Graph-Fusion-Experiment"
Repository = "https://github.com/sudheerdevu/Graph-Fusion-Experiment"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.black]
line-length = 100
